# [Enhancement] Fix performance problem of operator's mem tracker