--- a/be/src/runtime/current_thread.cpp
+++ b/be/src/runtime/current_thread.cpp
@@ -25,7 +25,7 @@ CurrentThread::~CurrentThread() {
         tls_is_thread_status_init = false;
         return;
     }
-    commit(true);
+    mem_tracker_ctx_shift();
     tls_is_thread_status_init = false;
 }
 
@@ -34,6 +34,10 @@ starrocks::MemTracker* CurrentThread::mem_tracker() {
         tls_mem_tracker = ExecEnv::GetInstance()->process_mem_tracker();
     }
     return tls_mem_tracker;
+}
+
+starrocks::MemTracker* CurrentThread::operator_mem_tracker() {
+    return tls_operator_mem_tracker;
 }
 
 CurrentThread& CurrentThread::current() {