--- a/be/src/runtime/current_thread.h
+++ b/be/src/runtime/current_thread.h
@@ -30,6 +30,9 @@
 #define SCOPED_THREAD_LOCAL_MEM_TRACKER_SETTER(mem_tracker) \
     auto VARNAME_LINENUM(tracker_setter) = CurrentThreadMemTrackerSetter(mem_tracker)
 
+#define SCOPED_THREAD_LOCAL_OPERATOR_MEM_TRACKER_SETTER(operator) \
+    auto VARNAME_LINENUM(tracker_setter) = CurrentThreadOperatorMemTrackerSetter(operator->mem_tracker())
+
 #define SCOPED_THREAD_LOCAL_CHECK_MEM_LIMIT_SETTER(check) \
     auto VARNAME_LINENUM(check_setter) = CurrentThreadCheckMemLimitSetter(check)
 
@@ -45,28 +48,104 @@ namespace starrocks {
 class TUniqueId;
 
 inline thread_local MemTracker* tls_mem_tracker = nullptr;
+inline thread_local MemTracker* tls_operator_mem_tracker = nullptr;
 inline thread_local MemTracker* tls_exceed_mem_tracker = nullptr;
 inline thread_local bool tls_is_thread_status_init = false;
 
 class CurrentThread {
-public:
-    CurrentThread() { tls_is_thread_status_init = true; }
-    ~CurrentThread();
+private:
+    class MemCacheManager {
+    public:
+        MemCacheManager(std::function<MemTracker*()>&& loader) : _loader(std::move(loader)) {}
+        MemCacheManager(const MemCacheManager&) = delete;
+        MemCacheManager(MemCacheManager&&) = delete;
 
-    void commit(bool is_ctx_shift) {
-        MemTracker* cur_tracker = mem_tracker();
-        if (cur_tracker != nullptr) {
-            cur_tracker->consume(_cache_size);
+        void consume(int64_t size) {
+            _cache_size += size;
+            _allocated_cache_size += size;
+            _total_consumed_bytes += size;
+            if (_cache_size >= BATCH_SIZE) {
+                commit(false);
+            }
+        }
+
+        bool try_mem_consume(int64_t size) {
+            MemTracker* cur_tracker = _loader();
+            _cache_size += size;
+            _allocated_cache_size += size;
+            _total_consumed_bytes += size;
+            if (cur_tracker != nullptr && _cache_size >= BATCH_SIZE) {
+                MemTracker* limit_tracker = cur_tracker->try_consume(_cache_size);
+                if (LIKELY(limit_tracker == nullptr)) {
+                    _cache_size = 0;
+                    return true;
+                } else {
+                    _cache_size -= size;
+                    _allocated_cache_size -= size;
+                    _try_consume_mem_size = size;
+                    tls_exceed_mem_tracker = limit_tracker;
+                    return false;
+                }
+            }
+            return true;
+        }
+
+        void release(int64_t size) {
+            _cache_size -= size;
+            _deallocated_cache_size += size;
+            if (_cache_size <= -BATCH_SIZE) {
+                commit(false);
+            }
+        }
+
+        void commit(bool is_ctx_shift) {
+            MemTracker* cur_tracker = _loader();
+            if (cur_tracker != nullptr) {
+                cur_tracker->consume(_cache_size);
+            }
             _cache_size = 0;
             if (is_ctx_shift) {
                 // Flush all cached info
-                cur_tracker->update_allocation(_allocated_cache_size);
-                cur_tracker->update_deallocation(_deallocated_cache_size);
+                if (cur_tracker != nullptr) {
+                    cur_tracker->update_allocation(_allocated_cache_size);
+                    cur_tracker->update_deallocation(_deallocated_cache_size);
+                }
                 _allocated_cache_size = 0;
                 _deallocated_cache_size = 0;
             }
         }
+
+        int64_t try_consume_mem_size() {
+            auto res = _try_consume_mem_size;
+            _try_consume_mem_size = 0;
+            return res;
+        }
+
+        int64_t get_consumed_bytes() const { return _total_consumed_bytes; }
+
+    private:
+        const static int64_t BATCH_SIZE = 2 * 1024 * 1024;
+
+        std::function<MemTracker*()> _loader;
+
+        // Allocated or delocated but not committed memory bytes, can be negative
+        int64_t _cache_size = 0;
+        // Allocated but not committed memory bytes, always positive
+        int64_t _allocated_cache_size = 0;
+        // Deallocated but not committed memory bytes, always positive
+        int64_t _deallocated_cache_size = 0;
+        int64_t _total_consumed_bytes = 0; // Totally consumed memory bytes
+        int64_t _try_consume_mem_size = 0; // Last time tried to consumed bytes
+    };
+
+public:
+    CurrentThread() : _mem_cache_manager(mem_tracker), _operator_mem_cache_manager(operator_mem_tracker) {
+        tls_is_thread_status_init = true;
     }
+    ~CurrentThread();
+
+    void mem_tracker_ctx_shift() { _mem_cache_manager.commit(true); }
+    void operator_mem_tracker_ctx_shift() { _operator_mem_cache_manager.commit(true); }
 
     void set_query_id(const starrocks::TUniqueId& query_id) { _query_id = query_id; }
     const starrocks::TUniqueId& query_id() { return _query_id; }
@@ -80,9 +159,17 @@ public:
 
     // Return prev memory tracker.
     starrocks::MemTracker* set_mem_tracker(starrocks::MemTracker* mem_tracker) {
-        commit(true);
+        mem_tracker_ctx_shift();
         auto* prev = tls_mem_tracker;
         tls_mem_tracker = mem_tracker;
+        return prev;
+    }
+
+    // Return prev memory tracker.
+    starrocks::MemTracker* set_operator_mem_tracker(starrocks::MemTracker* operator_mem_tracker) {
+        operator_mem_tracker_ctx_shift();
+        auto* prev = tls_operator_mem_tracker;
+        tls_operator_mem_tracker = operator_mem_tracker;
         return prev;
     }
 
@@ -95,6 +182,7 @@ public:
     bool check_mem_limit() { return _check; }
 
     static starrocks::MemTracker* mem_tracker();
+    static starrocks::MemTracker* operator_mem_tracker();
 
     static CurrentThread& current();
 
@@ -109,33 +197,21 @@ public:
     bool is_catched() const { return _is_catched; }
 
     void mem_consume(int64_t size) {
-        _cache_size += size;
-        _allocated_cache_size += size;
-        _total_consumed_bytes += size;
-        if (_cache_size >= BATCH_SIZE) {
-            commit(false);
-        }
+        _mem_cache_manager.consume(size);
+        _operator_mem_cache_manager.consume(size);
     }
 
     bool try_mem_consume(int64_t size) {
-        MemTracker* cur_tracker = mem_tracker();
-        _cache_size += size;
-        _allocated_cache_size += size;
-        _total_consumed_bytes += size;
-        if (cur_tracker != nullptr && _cache_size >= BATCH_SIZE) {
-            MemTracker* limit_tracker = cur_tracker->try_consume(_cache_size);
-            if (LIKELY(limit_tracker == nullptr)) {
-                _cache_size = 0;
-                return true;
-            } else {
-                _cache_size -= size;
-                _allocated_cache_size -= size;
-                _try_consume_mem_size = size;
-                tls_exceed_mem_tracker = limit_tracker;
-                return false;
-            }
+        if (_mem_cache_manager.try_mem_consume(size)) {
+            _operator_mem_cache_manager.consume(size);
+            return true;
         }
-        return true;
+        return false;
+    }
+
+    void mem_release(int64_t size) {
+        _mem_cache_manager.release(size);
+        _operator_mem_cache_manager.release(size);
     }
 
     static void mem_consume_without_cache(int64_t size) {
@@ -158,14 +234,6 @@ public:
         return true;
     }
 
-    void mem_release(int64_t size) {
-        _cache_size -= size;
-        _deallocated_cache_size += size;
-        if (_cache_size <= -BATCH_SIZE) {
-            commit(false);
-        }
-    }
-
     static void mem_release_without_cache(int64_t size) {
         MemTracker* cur_tracker = mem_tracker();
         if (cur_tracker != nullptr && size != 0) {
@@ -174,24 +242,18 @@ public:
     }
 
     // get last time try consume and reset
-    int64_t try_consume_mem_size() {
-        auto res = _try_consume_mem_size;
-        _try_consume_mem_size = 0;
-        return res;
-    }
+    int64_t try_consume_mem_size() { return _mem_cache_manager.try_consume_mem_size(); }
 
-    int64_t get_consumed_bytes() const { return _total_consumed_bytes; }
+    int64_t get_consumed_bytes() const { return _mem_cache_manager.get_consumed_bytes(); }
 
 private:
-    const static int64_t BATCH_SIZE = 2 * 1024 * 1024;
-    // Allocated or delocated but not committed memory bytes, can be negative
-    int64_t _cache_size = 0;
-    // Allocated but not committed memory bytes, always positive
-    int64_t _allocated_cache_size = 0;
-    // Deallocated but not committed memory bytes, always positive
-    int64_t _deallocated_cache_size = 0;
-    int64_t _total_consumed_bytes = 0; // Totally consumed memory bytes
-    int64_t _try_consume_mem_size = 0; // Last time tried to consumed bytes
+    // In order to record operator level memory trace while keep up high performance, we need to
+    // record the normal MemTracker's tree and operator's isolated MemTracker independently.
+    // `tls_operator_mem_tracker` will be updated every time when `Operator::pull_chunk` or `Operator::push_chunk`
+    // is invoked, the frequrency is a little bit high, but it does little harm to performance,
+    // because operator's MemTracker, which is a dangling MemTracker(withouth parent), has no concurrency conflicts
+    MemCacheManager _mem_cache_manager;
+    MemCacheManager _operator_mem_cache_manager;
     // Store in TLS for diagnose coredump easier
     TUniqueId _query_id;
     TUniqueId _fragment_instance_id;
@@ -221,6 +283,34 @@ public:
     CurrentThreadMemTrackerSetter(const CurrentThreadMemTrackerSetter&) = delete;
     void operator=(const CurrentThreadMemTrackerSetter&) = delete;
     CurrentThreadMemTrackerSetter(CurrentThreadMemTrackerSetter&&) = delete;
+    void operator=(CurrentThreadMemTrackerSetter&&) = delete;
+
+private:
+    MemTracker* _old_mem_tracker;
+    bool _is_same;
+};
+
+class CurrentThreadOperatorMemTrackerSetter {
+public:
+    explicit CurrentThreadOperatorMemTrackerSetter(MemTracker* new_mem_tracker) {
+        // operator's mem tracker must have no parent
+        DCHECK(new_mem_tracker == nullptr || new_mem_tracker->parent() == nullptr);
+        _old_mem_tracker = tls_thread_status.operator_mem_tracker();
+        _is_same = (_old_mem_tracker == new_mem_tracker);
+        if (!_is_same) {
+            tls_thread_status.set_operator_mem_tracker(new_mem_tracker);
+        }
+    }
+
+    ~CurrentThreadOperatorMemTrackerSetter() {
+        if (!_is_same) {
+            (void)tls_thread_status.set_operator_mem_tracker(_old_mem_tracker);
+        }
+    }
+
+    CurrentThreadOperatorMemTrackerSetter(const CurrentThreadOperatorMemTrackerSetter&) = delete;
+    void operator=(const CurrentThreadOperatorMemTrackerSetter&) = delete;
+    CurrentThreadOperatorMemTrackerSetter(CurrentThreadOperatorMemTrackerSetter&&) = delete;
     void operator=(CurrentThreadMemTrackerSetter&&) = delete;
 
 private: