@wanpengfei-git 2023-02-27
run starrocks_admit_test